# gh-95385 Fastpath for encoding dict to JSON