--- a/Misc/NEWS.d/next/Library/2022-07-28-17-14-38.gh-issue-95385.6YlsDI.rst
+++ b/Misc/NEWS.d/next/Library/2022-07-28-17-14-38.gh-issue-95385.6YlsDI.rst
@@ -1,0 +1,1 @@
+Faster ``json.dumps()`` when sorting of keys is not requested (default).