--- a/Modules/_json.c
+++ b/Modules/_json.c
@@ -1593,7 +1593,6 @@ encoder_listencode_dict(PyEncoderObject *s, _PyUnicodeWriter *writer,
     }
 
     if (s->sort_keys) {
-        Py_ssize_t i;
 
         items = PyDict_Items(dct);
         if (items == NULL || PyList_Sort(items) < 0)