When sorting of keys is not requested and we are encoding a dict
use PyDict_Next to iterate over keys and values.
When sorting is requested use PyList_GET_ITEM instead of PyIter_Next because we know we are working with a list.

<!-- gh-issue-number: gh-95385 -->
* Issue: gh-95385
<!-- /gh-issue-number -->