@aivarsk 2022-07-28
I have made the requested changes; please review again