@aivarsk 2022-07-29
pyperformance on my machine before changes:

```
### json_dumps ###
Mean +- std dev: 10.6 ms +- 0.1 ms
```

pyperformance after changes:

```
### json_dumps ###
Mean +- std dev: 9.00 ms +- 0.08 ms
```