@aivarsk 2022-07-29
With pyperformance sort_keys=True

before changes
```
### json_dumps ###
Mean +- std dev: 14.0 ms +- 0.2 ms
```
after changes (PyList_GET_ITEM is a bit faster than PyIter_Next?)
```
### json_dumps ###
Mean +- std dev: 13.6 ms +- 0.2 ms
```